update: f user管理画面で給与の設定を確認機能追加

+++ front-end/pages/admin/user/index.vue
@@ -17,7 +17,52 @@
           delete
         </v-btn>
       </template>
+      <template v-slot:[`item.salary`]="{ item }">
+        <v-btn v-if="item.salary == null" small color="error" @click="openDialog(item)">給与未設定</v-btn>
+        <v-btn v-if="item.salary != null" small color="info">給与設定済み</v-btn>
+      </template>
     </v-data-table>
+      <!-- ダイアログ -->
+      <v-dialog
+        v-model="dialog"
+        max-width="500px"
+        transition="dialog-transition"
+      >
+        <v-card>
+          <v-card-title primary-title>
+            {{ form_data.name }}さんの給与設定
+          </v-card-title>
+          <v-card-text>
+            <v-text-field
+              name="user_id"
+              type="hidden"
+              v-model="form_data.user_id"
+            ></v-text-field>
+            <v-select
+              :items="select_items"
+              v-model="form_data.salary_type"
+              item-text="text"
+              item-value="value"
+              label="給与のタイプ"
+            ></v-select>
+            <v-text-field
+              name="hour_salary"
+              label="時給"
+              type="number"
+              v-model="form_data.hour_salary"
+            ></v-text-field>
+            <v-text-field
+              name="month_salary"
+              label="月給"
+              type="number"
+              v-model="form_data.month_salary"
+            ></v-text-field>
+          </v-card-text>
+          <v-card-actions>
+            <v-btn color="success" @click="addSalary">決定</v-btn>
+          </v-card-actions>
+        </v-card>
+      </v-dialog>
   </div>
 </template>
 
@@ -26,6 +71,8 @@ export default {
   data() {
     return {
       is_load: false,
+      dialog: false,
+      form_data: {},
       headers: [
         {
           text: 'ID',
@@ -43,9 +90,18 @@ export default {
           text: '削除',
           value: 'delete',
           sortable: false,
+        },
+        {
+          text: '給与',
+          value: 'salary',
+          sortable: false,
         }
       ],
       items: [],
+      select_items: [
+        { text: '時給', value: 'hour' },
+        { text: '月給', value: 'month' }
+      ],
     }
   },
   methods: {
@@ -62,7 +118,11 @@ export default {
       });
     },
     async deleteItem(item) {
-      confirm('本当に削除しますか')
+      let res = confirm('本当に削除しますか')
+
+      if (res == false) {
+        return;
+      }
 
       await this.$axios.delete('/api/admin/user', {
         data: {
@@ -77,8 +137,20 @@ export default {
       .catch(e => {
         this.$store.dispatch('flashMessage/showErrorMessage', '削除に失敗しました');
       })
-
     },
+    openDialog(item) {
+      this.dialog = true;
+      this.form_data.name = item.name;
+      this.form_data.user_id = item.id;
+      if(item.salary) {
+        this.form_data.salary_type = item.salary.salary_type;
+        this.form_data.hour_salary = item.salary.hour_salary;
+        this.form_data.month_salary = item.salary.month_salary;
+      }
+    },
+    addSalary() {
+      console.log(this.form_data);
+    }
   },
   mounted() {
     this.getUser();
